Started work on decouplig and setting the env for testing
