+++ organise_desktop/Extension.json
@@ -1,14 +1,14 @@
 {
 	"Shortcuts":
-	[".kys",".lnk",".mat",".pif",".shb",".url","xnk"], 
-	
+	[".kys",".lnk",".mat",".pif",".shb",".url","xnk"],
+
 	"Executables":
-	[".action",".apk",".app",".bin",".eham",".exe",".ham",".msi",".osx",".out",".pex",".plx",".prg",".run",".scr",".upx"], 
+	[".action",".apk",".app",".bin",".eham",".exe",".ham",".msi",".osx",".out",".pex",".plx",".prg",".run",".scr",".upx"],
 
 	"Malware":
 	[".APPLICATION",".GADGET",".CPL",".MSC",".JS",".JSE",".WS","WSF",".REG",".CSH"],
-	
-	
+
+
 	"Folders":
 	[".0",".000",".7z",".7z.001",".7z.002",".a00",".a01",".a02",".ace",".agg",".ain",".alz",".apz",".ar",".arc",".archiver",".arduboy",".arh",
 	".ari",".arj",".ark",".asr",".b1",".b64",".ba",".bdoc",".bh",".bndl",".boo",".bundle",".bz",".bz2",".bza",".bzip",".bzip2",".c00",".c01",
@@ -24,16 +24,16 @@
 	".stproj",".sy_",".tar.bz2",".tar.gz",".tar.gz2",".tar.lz",".tar.lzma",".tar.xz",".tar.z",".taz",".tbz",".tbz2",".tg",".tgz",".tlz",
 	".tlzma",".trs",".tx_",".txz",".tz",".uc2",".ufs.uzip",".uha",".uzip",".vem",".vip",".voca",".vsi",".wa",".waff",".war",".wlb",".wot",
 	".xapk",".xar",".xef",".xez",".xip",".xmcdz",".xx",".xz",".xzm",".y",".yz",".yz1",".z",".z01",".z02",".z03",".z04",".zap",".zfsendtotarget",
-	".zi",".zip",".zipx",".zix",".zl",".zoo",".zpi",".zsplit",".zw",".zz"], 
-	
+	".zi",".zip",".zipx",".zix",".zl",".zoo",".zpi",".zsplit",".zw",".zz"],
+
 	"Images":
 	[".3fr",".ai",".ari",".arr",".arw",".avatar",".awd",".bmp",".bpg",".cdmz",".cdr",".cr2",".crw",".ct",".dcr",".dds",
 	".ded",".djvu",".dng",".eip",".eps",".exr",".fff",".fpx",".gif",".hdr",".hpgl",".i3d",".ico",".icon",".itc2",".iwi",
 	".j2c",".j2k",".jbig2",".jfif",".jif",".jp2",".jpc",".jpe",".jpeg",".jpg",".jpg-large",".jpx",".jxr",".k25",".kdc",
 	".lip",".mfw",".mos",".msp",".nef",".nrw",".orf",".ota",".pat",".pcd",".pcx",".pdn",".pi2",".pic",".picnc",".piskel",
 	".png",".ppf",".psb",".psd",".pspbrush",".pwp",".pxd",".pxr",".raf",".raw",".rw2",".rwz",".sai",".spr",".sprite",
-	".sprite2",".srf",".sumo",".svg",".tbn",".tga",".tif",".tiff",".tn",".tn3",".ufo",".vsdx",".webp",".x3f",".xcf",".xpm"], 
-	
+	".sprite2",".srf",".sumo",".svg",".tbn",".tga",".tif",".tiff",".tn",".tn3",".ufo",".vsdx",".webp",".x3f",".xcf",".xpm"],
+
 	"Music":
 	[".aiff", ".aac", ".alac", ".flp", ".mp3", ".aimppl", ".aup", ".ang", ".mscz", ".vdj", ".ram", ".sf2", ".sfk", ".xfs",
 	".dff", ".als", ".bidule", ".bnk", ".bww", ".cgrp", ".csh", ".cwb", ".dfc", ".dig", ".dm", ".dmsa", ".dmse", ".dsf",
@@ -42,8 +42,8 @@
 	".sou", ".sprg", ".sseq", ".sxt", ".syn", ".syw", ".trak", ".ust", ".vsqx", ".w64", ".wus", ".zpa", ".zvd", ".flac", ".afc",
 	".mui", ".abm", ".pla", ".vag", ".act", ".wav", ".bun", ".gsm", ".midi", ".m3u", ".mmm", ".pkf", ".sng", ".m4r", ".mpga",
 	".omf", ".rx2", ".flm", ".gp5", ".sdt", ".ac3", ".mid", ".xspf", ".nkm", ".nsa", ".ply", ".sbg", ".sdat", ".toc", ".tg",
-	".ogg", ".wma", ".amr", ".m3u8", ".5xe", ".aif", ".frg", ".minigsf", ".pcast", ".vyf"], 
-	
+	".ogg", ".wma", ".amr", ".m3u8", ".5xe", ".aif", ".frg", ".minigsf", ".pcast", ".vyf"],
+
 	"Movies":
 	[".264",".3g2",".3gp",".3gp2",".3gpp",".3gpp2",".3mm",".3p2",".60d",".787",".890",".aaf",".aec",".aecap",".aegraphic",".aep",
 	".aepx",".aet",".aetx",".ajp",".ale",".am",".amc",".amv",".amx",".anim",".anx",".aqt",".arcut",".arf",".asf",".asx",".avb",
@@ -73,8 +73,8 @@
 	".vep",".vf",".vft",".vfw",".vfz",".vgz",".vid",".video",".viewlet",".viv",".vivo",".vix",".vlab",".vmlf",".vmlt",".vob",".vp3",
 	".vp6",".vp7",".vpj",".vr",".vro",".vs4",".vse",".vsp",".vtt",".w32",".wcp",".webm",".wfsp",".wgi",".wlmp",".wm",".wmd",".wmmp",
 	".wmv",".wmx",".wot",".wp3",".wpl",".wsve",".wtv",".wve",".wvm",".wvx",".wxp",".xej",".xel",".xesc",".xfl",".xlmv",".xml",".xmv",
-	".xvid",".y4m",".yog",".yuv",".zeg",".zm1",".zm2",".zm3",".zmv"], 
-	
+	".xvid",".y4m",".yog",".yuv",".zeg",".zm1",".zm2",".zm3",".zmv"],
+
 	"Text":
 	[ ".1st",".abw",".accdb",".act",".adoc",".aim",".ans",".apkg",".apt",".arff",".asc",".ascii",".ase",".aty",".awp",".awt",".aww",".bad",
 	".bbs",".bdp",".bdr",".bean",".bib",".bibtex",".bml",".bna",".boc",".brx",".btd",".bzabw",".c",".calca",".cdz",".charset",".chart",
@@ -92,12 +92,12 @@
 	".story",".strings",".stw",".sty",".sub",".sublime-project",".sublime-workspace",".sxg",".sxw",".tab",".tdf",".template",".tex",
 	".text",".textclipping",".thp",".tlb",".tm",".tmd",".tmv",".tpc",".trelby",".tvj",".txt",".u3i",".unauth",".unx",".uof",".uot",
 	".upd",".utf8",".utxt",".vct",".vnt",".vw",".wbk",".webdoc",".wn",".wp",".wp4",".wp5",".wp6",".wp7",".wpa",".wpd",".wpl",".wps",
-	".wpt",".wpw",".wri",".wsd",".wtt",".wtx",".xbdoc",".xbplate",".xdl",".xlsx",".xwp",".xy",".xy3",".xyp",".xyw",".zabw",".zrtf",".zw"], 
-	
+	".wpt",".wpw",".wri",".wsd",".wtt",".wtx",".xbdoc",".xbplate",".xdl",".xlsx",".xwp",".xy",".xy3",".xyp",".xyw",".zabw",".zrtf",".zw", "xls"], 
+
 	"CAD":
 	[".ma", ".dxe", ".gpw", ".mb", ".mnx", ".shx", ".dwk", ".apj", ".sld", ".hdi", ".dwg", ".lli", ".aws", ".dwz", ".slb", ".dxx", ".dw2l",
 		".dwt", ".dws", ".shp", ".dbt", ".dxf", ".schematic", ".pwt", ".mvi", ".blk", ".fbx", ".blend", ".chc"],
-	
+
 	"Programming":
 	[".gmo", ".dcu", ".wixmst", ".csproj", ".r", ".pjx", ".lgo", ".bcp", ".vbp", ".vdm", ".diff", ".mv", ".pyw", ".ilk", ".vdproj", ".mdzip",
 		".bpg", ".resx", ".agi", ".gmd", ".proto", ".resjson", ".pri", ".gorm", ".f", ".vtv", ".xcappdata", ".groupproj", ".mf", ".sln", ".ccn",
