+++ organise_desktop/Read_Me.txt
@@ -2,27 +2,14 @@ This is a list of all the items on your desktop before it was cleaned.
 Email this list to kalimbatech@gmail.com if anything is not working as planned, it will help with debugging
 Together we can make a better app
 
-untitled folder
 Music
-ar.pdf
-Finished._files
-Screenshots
-.DS_Store
-LoL Logs.zip
-.localized
-edf.pdf
 Images
 Executables
-6QKLA16Y5T.txt
-edf-nov-dec.pdf
-acte.pdf
 Malware
 CAD
 Programming
-Finished..html
 Shortcuts
 Movies
 Folders
 Text
-floatingpointspec.pdf
-traseu.txt
+OrgDesk.xls
