+++ organise_desktop/undo.py
@@ -20,8 +20,10 @@ def execute():
 
         '''list of folders to be created'''
     elif sys.platform == 'linux' or sys.platform == 'darwin':
-        desktopdir = path.join(environ['HOME'], 'Desktop')
-
+        if environ['TEST_DIR'] != '':
+            desktopdir = environ['TEST_DIR']
+        else:
+            desktopdir = path.join(environ['HOME'], 'Desktop')
     else:
         print("{} version not implemented".format(sys.platform))
         raise NotImplementedError
@@ -37,5 +39,3 @@ def execute():
             for thing in contents:
                 rename(src=desktopdir+separator+folder+separator+thing, dst=desktopdir+separator+thing)
             rmdir(desktopdir+separator+folder)
-
-
